refactor: moved github files to github dir
