+++ lib/github/get-gist.js
@@ -1,9 +1,9 @@
 const got = require('got');
-const cacheFallback = require('./utils/cache-fallback');
-const githubApi = require('./utils/github-api');
-const { github } = require('./utils/config');
-const resError = require('./utils/res-error');
-const config = require('./utils/config');
+const cacheFallback = require('../utils/cache-fallback');
+const githubApi = require('../utils/github-api');
+const { github } = require('../utils/config');
+const resError = require('../utils/res-error');
+const config = require('../utils/config');
 
 const cacheDuration = 10000;
