+++ lib/github/get-path.js
@@ -1,9 +1,9 @@
 const qs = require('querystring');
 const nodePath = require('path');
-const cacheFallback = require('./utils/cache-fallback');
-const githubApi = require('./utils/github-api');
-const { github } = require('./utils/config');
-const resError = require('./utils/res-error');
+const cacheFallback = require('../utils/cache-fallback');
+const githubApi = require('../utils/github-api');
+const { github } = require('../utils/config');
+const resError = require('../utils/res-error');
 
 
 const cacheDuration = 10000;
