+++ lib/github/get-remote-info.js
@@ -1,8 +1,8 @@
 const { getRemoteInfo: igGetRemoteInfo } = require('isomorphic-git');
 const http = require('isomorphic-git/http/node');
-const { github } = require('./utils/config');
-const cacheFallback = require('./utils/cache-fallback');
-const resError = require('./utils/res-error');
+const { github } = require('../utils/config');
+const cacheFallback = require('../utils/cache-fallback');
+const resError = require('../utils/res-error');
 
 const cacheDuration = 10000;
