+++ lib/resolve-ref.js
@@ -1,5 +1,5 @@
 const semver = require('semver');
-const getRemoteInfo = require('./github.get-remote-info');
+const getRemoteInfo = require('./github/get-remote-info');
 
 async function resolveRef(args) {
 	let { ref } = args;
